Check in a canonical PKI
For some reason initpki doesn't generate a set of certs that are
acceptable to the P-Ras core, even though I never modified it.
Maybe some defaults changed in a recent release of OpenSSL or
something.

Well, to make everybody's life easier let's all just use the same
private keys for everything. It's not like it matters.